Move function off of parent onto FacetItem
