+++ src/components/Search/SearchFacets.jsx
@@ -2,7 +2,6 @@
 var React = require('react');
 var mui = require('material-ui');
 var List = mui.List;
-var SearchActions = require('../../actions/SearchActions.js');
 var SearchStore = require('../../store/SearchStore.js');
 var FacetItem = require('./FacetItem.jsx');
 
@@ -10,19 +9,6 @@ var SearchFacets = React.createClass({
   mixins: [
     require('../../mixins/CurrentThemeMixin.jsx')
   ],
-
-  valueOnClick: function(e) {
-    var values = e.currentTarget.getAttribute("value").split("|");
-    if(SearchStore.facetOption) {
-      for(var i = 0; i < SearchStore.facetOption.length; i++) {
-        if (SearchStore.facetOption[i].name === values[0] && SearchStore.facetOption[i].value === values[1]) {
-          SearchStore.removeSelectedFacet({ name: values[0], value: values[1] });
-          return
-        }
-      }
-    }
-    SearchActions.setSelectedFacet({ name: values[0], value: values[1] });
-  },
 
   values: function(facet) {
     var parentFacet = facet.field;
@@ -40,7 +26,6 @@ var SearchFacets = React.createClass({
           <FacetItem
             field={ facet.field}
             facet={ e }
-            clickAction={ this.valueOnClick }
             key={ e.name }
           />
         );
